refurbish bad decisions
